tipsy-account/src/main/java/com/ssafy/tipsy/auth/filter/JwtAuthenticationEntryPoint.java
package com.ssafy.tipsy.auth.filter;

import com.ssafy.tipsy.auth.model.AuthFail;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.AuthenticationEntryPoint;
import org.springframework.stereotype.Component;

import java.io.IOException;

@Component
public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
    @Override
    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
        Object exception = request.getAttribute("Exception");
        if(exception instanceof AuthFail authFail) {
            response.setStatus((authFail).getCode());
            response.setCharacterEncoding("utf-8");
            response.setContentType("text/html; charset=UTF-8");
            response.getWriter().write(authFail.getMsg());
        }
    }
}